Refactoring rental card component
