+++ ecommerce-ui/src/UI/ListItems.jsx
@@ -5,31 +5,36 @@ import IconButton from "@mui/material/IconButton";
 import ListItem from "@mui/material/ListItem";
 import ListItemText from "@mui/material/ListItemText";
 import PropTypes from "prop-types";
+import { Fade } from "@mui/material";
 
 export default function ListItems({ title, fee, city, length }) {
   const rentalCart = useContext(RentalCartContext);
 
   const onClickHandler = () => {
-    if(window.confirm(`Are you sure you want to remove the following rental from your cart? 
-    *** ${title} ***`)) {
+    if (
+      window.confirm(`Are you sure you want to remove the following rental from your cart? 
+    *** ${title} ***`)
+    ) {
       rentalCart.removeRental(title);
-    } else alert("No changes were made to your cart.")
+    } else alert("No changes were made to your cart.");
   };
 
   return (
     <>
-      <ListItem
-        secondaryAction={
-          <IconButton edge="end" aria-label="delete" onClick={onClickHandler}>
-            <DeleteIcon />
-          </IconButton>
-        }
-      >
-        <ListItemText
-          primary={`${title} - ${city}`}
-          secondary={`$${fee}/day, Days requested: ${length} `}
-        />
-      </ListItem>
+      <Fade in={true} timeout={500}>
+        <ListItem
+          secondaryAction={
+            <IconButton edge="end" aria-label="delete" onClick={onClickHandler}>
+              <DeleteIcon />
+            </IconButton>
+          }
+        >
+          <ListItemText
+            primary={`${title} - ${city}`}
+            secondary={`$${fee}/day, Days requested: ${length} `}
+          />
+        </ListItem>
+      </Fade>
     </>
   );
 }
